Refactor cluster commands into new package
This is a larger command and it made sense to break it out given we're adding a bunch more
