+++ internal/command/users.go
@@ -11,6 +11,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/jetstack/jsctl/internal/client"
+	errors2 "github.com/jetstack/jsctl/internal/command/errors"
 	"github.com/jetstack/jsctl/internal/config"
 	"github.com/jetstack/jsctl/internal/prompt"
 	"github.com/jetstack/jsctl/internal/table"
@@ -45,7 +46,7 @@ func usersList() *cobra.Command {
 			http := client.New(ctx, apiURL)
 			cnf, ok := config.FromContext(ctx)
 			if !ok || cnf.Organization == "" {
-				return errNoOrganizationName
+				return errors2.ErrNoOrganizationName
 			}
 
 			users, err := user.List(ctx, http, cnf.Organization)
@@ -89,7 +90,7 @@ func usersAdd() *cobra.Command {
 			http := client.New(ctx, apiURL)
 			cnf, ok := config.FromContext(ctx)
 			if !ok || cnf.Organization == "" {
-				return errNoOrganizationName
+				return errors2.ErrNoOrganizationName
 			}
 
 			email := args[0]
@@ -123,7 +124,7 @@ func usersRemove() *cobra.Command {
 			http := client.New(ctx, apiURL)
 			cnf, ok := config.FromContext(ctx)
 			if !ok || cnf.Organization == "" {
-				return errNoOrganizationName
+				return errors2.ErrNoOrganizationName
 			}
 
 			email := args[0]
